+++ tests/testthat/tests.make_stancode.R
@@ -1340,12 +1340,16 @@ test_that("Stan code for CAR models is correct", {
   expect_match2(scode, "target += sparse_car_lpdf(")
   expect_match2(scode, "mu[n] += rcar[Jloc[n]]")
 
-  scode <- make_stancode(y ~ x, dat, autocor = cor_icar(W),
-                         silent = TRUE)
+  scode <- make_stancode(y ~ x, dat, autocor = cor_car(W, type = "esicar"))
   expect_match2(scode, "real sparse_icar_lpdf(vector phi")
   expect_match2(scode, "target += sparse_icar_lpdf(")
   expect_match2(scode, "mu[n] += rcar[Jloc[n]]")
   expect_match2(scode, "rcar[Nloc] = - sum(zcar)")
+  
+  scode <- make_stancode(y ~ x, dat, autocor = cor_icar(W))
+  expect_match2(scode, "target += -0.5 * dot_self(rcar[edges1] - rcar[edges2])")
+  expect_match2(scode, "target += normal_lpdf(sum(rcar) | 0, 0.001 * Nloc)")
+  expect_match2(scode, "mu[n] += rcar[Jloc[n]]")
 })
 
 test_that("Stan code for skew_normal models is correct", {